move crash reporting in a separate file
